WIP Start discecting angular doc implementation
At this point the docs are generating based on the project data but still
using all of the branding/images/etc. of angular. Since the dgeni lib
 is still a bit lacking I found it easiest to essentially copy/paste parts
 of the AngularJS source to get a base framework in place for providing
 a minimal but functional implementation.

The next step is to strip out everything specific to Angular and either
contribute to the dgeni project or extract it into it's own if it doesn't
fit.